admin id unhardcoded and Core namespace renamed to BotCore
